@@ -1,7 +1,7 @@
 {
   "author": "Auth0",
   "name": "@auth0/auth0-react",
-  "version": "1.12.1",
+  "version": "2.0.0-beta.0",
   "description": "Auth0 SDK for React Single Page Applications (SPA)",
   "keywords": [
     "auth0",
@@ -31,7 +31,7 @@
     "start:gatsby": "npm start --prefix=examples/gatsby-app",
     "start:nextjs": "npm run dev --prefix=examples/nextjs-app",
     "start:api": "npm start --prefix=examples/users-api",
-    "test:cra": "start-server-and-test start:api 3001 start:cra 3000 cypress:run",
+    "test:cra": "start-server-and-test start:api 3001 start:cra http-get://127.0.0.1:3000 cypress:run",
     "test:cra:watch": "start-server-and-test start:api 3001 start:cra 3000 cypress:open",
     "test:gatsby": "start-server-and-test start:api 3001 start:gatsby 3000 cypress:run",
     "test:gatsby:watch": "start-server-and-test start:api 3001 start:gatsby 3000 cypress:open",
@@ -51,8 +51,9 @@
   },
   "homepage": "https://github.com/auth0/auth0-react#readme",
   "devDependencies": {
-    "@rollup/plugin-node-resolve": "^13.1.3",
-    "@rollup/plugin-replace": "^4.0.0",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-replace": "^5.0.1",
+    "@rollup/plugin-terser": "^0.1.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/react-hooks": "^7.0.2",
@@ -61,7 +62,7 @@
     "@types/react-dom": "^17.0.18",
     "@typescript-eslint/eslint-plugin": "^5.45.0",
     "@typescript-eslint/parser": "^5.45.0",
-    "browserstack-cypress-cli": "1.8.1",
+    "browserstack-cypress-cli": "^1.19.1",
     "cypress": "^9.7.0",
     "eslint": "^8.28.0",
     "eslint-plugin-react": "^7.31.11",
@@ -70,26 +71,25 @@
     "jest": "^29.3.1",
     "jest-environment-jsdom": "^29.3.1",
     "jest-junit": "^15.0.0",
-    "oidc-provider": "^7.10.6",
-    "prettier": "2.8.0",
+    "oidc-provider": "^8.0.0",
+    "prettier": "^2.8.1",
     "pretty-quick": "^3.1.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-test-renderer": "^18.2.0",
-    "rollup": "^2.70.1",
+    "rollup": "^3.7.0",
     "rollup-plugin-analyzer": "^4.0.0",
     "rollup-plugin-delete": "^2.0.0",
     "rollup-plugin-dev": "^1.1.3",
     "rollup-plugin-livereload": "^2.0.5",
     "rollup-plugin-peer-deps-external": "^2.2.4",
-    "rollup-plugin-serve": "^1.1.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.31.2",
-    "start-server-and-test": "^1.14.0",
+    "rollup-plugin-serve": "^2.0.2",
+    "rollup-plugin-typescript2": "^0.34.1",
+    "start-server-and-test": "^1.15.2",
     "ts-jest": "^29.0.3",
-    "tslib": "^2.3.1",
-    "typedoc": "^0.22.13",
-    "typescript": "^4.6.3"
+    "tslib": "^2.4.1",
+    "typedoc": "^0.23.21",
+    "typescript": "^4.9.4"
   },
   "peerDependencies": {
     "react": "^16.11.0 || ^17 || ^18",
@@ -101,6 +101,6 @@
     }
   },
   "dependencies": {
-    "@auth0/auth0-spa-js": "^1.22.6"
+    "@auth0/auth0-spa-js": "^2.0.2"
   }
 }
