@@ -2,11 +2,11 @@ import '@testing-library/jest-dom/extend-expect';
 import React from 'react';
 import withAuthenticationRequired from '../src/with-authentication-required';
 import { render, screen, waitFor, act } from '@testing-library/react';
-import { Auth0Client, User } from '@auth0/auth0-spa-js';
+import { Auth0Client} from '@auth0/auth0-spa-js';
 import Auth0Provider from '../src/auth0-provider';
 import { Auth0ContextInterface, initialContext } from '../src/auth0-context';
 
-const mockClient = jest.mocked(new Auth0Client({ client_id: '', domain: '' }));
+const mockClient = jest.mocked(new Auth0Client({ clientId: '', domain: '' }));
 
 describe('withAuthenticationRequired', () => {
   it('should block access to a private component when not authenticated', async () => {
@@ -28,66 +28,6 @@ describe('withAuthenticationRequired', () => {
     mockClient.getUser.mockResolvedValue({ name: '__test_user__' });
     const MyComponent = (): JSX.Element => <>Private</>;
     const WrappedComponent = withAuthenticationRequired(MyComponent);
-    render(
-      <Auth0Provider clientId="__test_client_id__" domain="__test_domain__">
-        <WrappedComponent />
-      </Auth0Provider>
-    );
-    await waitFor(() =>
-      expect(mockClient.loginWithRedirect).not.toHaveBeenCalled()
-    );
-    await waitFor(() =>
-      expect(screen.getByText('Private')).toBeInTheDocument()
-    );
-  });
-
-  it('should not allow access to claims-restricted components', async () => {
-    const MyComponent = (): JSX.Element => <>Private</>;
-    const WrappedComponent = withAuthenticationRequired(MyComponent, {
-      claimCheck: (claims?: User) =>
-        claims?.['https://my.app.io/jwt/roles']?.includes('ADMIN'),
-    });
-    /**
-     * A user with USER and MODERATOR roles.
-     */
-    const mockUser = {
-      name: '__test_user__',
-      'https://my.app.io/jwt/claims': {
-        USER: '__test_user__',
-        ROLE: ['USER', 'MODERATOR'],
-      },
-    };
-    mockClient.getUser.mockResolvedValue(mockUser);
-
-    render(
-      <Auth0Provider clientId="__test_client_id__" domain="__test_domain__">
-        <WrappedComponent />
-      </Auth0Provider>
-    );
-    await waitFor(() =>
-      expect(mockClient.loginWithRedirect).toHaveBeenCalled()
-    );
-    expect(screen.queryByText('Private')).not.toBeInTheDocument();
-  });
-
-  it('should allow access to restricted components when JWT claims present', async () => {
-    const MyComponent = (): JSX.Element => <>Private</>;
-    const WrappedComponent = withAuthenticationRequired(MyComponent, {
-      claimCheck: (claim?: User) =>
-        claim?.['https://my.app.io/jwt/claims']?.ROLE?.includes('ADMIN'),
-    });
-    /**
-     * User with ADMIN role.
-     */
-    const mockUser = {
-      name: '__test_user__',
-      'https://my.app.io/jwt/claims': {
-        USER: '__test_user__',
-        ROLE: ['ADMIN'],
-      },
-    };
-    mockClient.getUser.mockResolvedValue(mockUser);
-
     render(
       <Auth0Provider clientId="__test_client_id__" domain="__test_domain__">
         <WrappedComponent />
