@@ -6,13 +6,14 @@ import { Error } from './Error';
 const PORT = process.env.REACT_APP_API_PORT || 3001;
 
 export function Users() {
-  const { loading, error, data: users = [] } = useApi(
-    `http://localhost:${PORT}/users`,
-    {
-      audience: process.env.REACT_APP_AUDIENCE,
-      scope: 'read:users',
-    }
-  );
+  const {
+    loading,
+    error,
+    data: users = [],
+  } = useApi(`http://localhost:${PORT}/users`, {
+    audience: process.env.REACT_APP_AUDIENCE,
+    scope: 'profile email read:users',
+  });
 
   if (loading) {
     return <Loading />;
