@@ -16,7 +16,9 @@ export const useApi = (
     (async () => {
       try {
         const { audience, scope, ...fetchOptions } = options;
-        const accessToken = await getAccessTokenSilently({ audience, scope });
+        const accessToken = await getAccessTokenSilently({
+          authorizationParams: { audience, scope },
+        });
         const res = await fetch(url, {
           ...fetchOptions,
           headers: {
