@@ -1,8 +1,3 @@
 module.exports = {
-  plugins: [
-    {
-      resolve: 'gatsby-plugin-create-client-paths',
-      options: { prefixes: ['/*'] },
-    },
-  ],
+  trailingSlash: 'never',
 };
