@@ -4,7 +4,7 @@ export function Loading() {
   return (
     <div className="text-center">
       <div className="spinner-border" role="status">
-        <span className="sr-only">Loading...</span>
+        <span className="visually-hidden">Loading...</span>
       </div>
     </div>
   );
