@@ -7,47 +7,53 @@ export function Nav() {
   const pathname = typeof window !== 'undefined' && window.location.pathname;
 
   return (
-    <nav className="navbar navbar-expand-lg navbar-light bg-light">
-      <span className="navbar-brand">@auth0/auth0-react</span>
-      <div className="collapse navbar-collapse">
-        <div className="navbar-nav">
-          <Link
-            to="/"
-            className={`nav-item nav-link${pathname === '/' ? ' active' : ''}`}
-          >
-            Home
-          </Link>
-          <Link
-            to="/users"
-            className={`nav-item nav-link${
-              pathname === '/users' ? ' active' : ''
-            }`}
-          >
-            Users
-          </Link>
+    <nav className="navbar navbar-expand-md navbar-light bg-light">
+      <div className="container-fluid">
+        <span className="navbar-brand">@auth0/auth0-react</span>
+        <div className="collapse navbar-collapse">
+          <div className="navbar-nav">
+            <Link
+              to="/"
+              className={`nav-item nav-link${
+                pathname === '/' ? ' active' : ''
+              }`}
+            >
+              Home
+            </Link>
+            <Link
+              to="/users"
+              className={`nav-item nav-link${
+                pathname === '/users' ? ' active' : ''
+              }`}
+            >
+              Users
+            </Link>
+          </div>
         </div>
-      </div>
 
-      {isAuthenticated ? (
-        <div>
-          <span id="hello">Hello, {user.name}!</span>{' '}
+        {isAuthenticated ? (
+          <div>
+            <span id="hello">Hello, {user.name}!</span>{' '}
+            <button
+              className="btn btn-outline-secondary"
+              id="logout"
+              onClick={() =>
+                logout({ logoutParams: { returnTo: window.location.origin } })
+              }
+            >
+              logout
+            </button>
+          </div>
+        ) : (
           <button
-            className="btn btn-outline-secondary"
-            id="logout"
-            onClick={() => logout({ returnTo: window.location.origin })}
+            className="btn btn-outline-success"
+            id="login"
+            onClick={() => loginWithRedirect()}
           >
-            logout
+            login
           </button>
-        </div>
-      ) : (
-        <button
-          className="btn btn-outline-success"
-          id="login"
-          onClick={() => loginWithRedirect()}
-        >
-          login
-        </button>
-      )}
+        )}
+      </div>
     </nav>
   );
 }
