@@ -5,13 +5,8 @@ import { useRouter } from 'next/router';
 import { Loading } from './Loading';
 
 export function Nav() {
-  const {
-    isAuthenticated,
-    isLoading,
-    user,
-    loginWithRedirect,
-    logout,
-  } = useAuth0();
+  const { isAuthenticated, isLoading, user, loginWithRedirect, logout } =
+    useAuth0();
   const { pathname } = useRouter();
 
   if (isLoading) {
@@ -23,23 +18,19 @@ export function Nav() {
       <span className="navbar-brand">@auth0/auth0-react</span>
       <div className="collapse navbar-collapse">
         <div className="navbar-nav">
-          <Link href="/">
-            <a
-              className={`nav-item nav-link${
-                pathname === '/' ? ' active' : ''
-              }`}
-            >
-              Home
-            </a>
+          <Link
+            href="/"
+            className={`nav-item nav-link${pathname === '/' ? ' active' : ''}`}
+          >
+            Home
           </Link>
-          <Link href="/users">
-            <a
-              className={`nav-item nav-link${
-                pathname === '/users' ? ' active' : ''
-              }`}
-            >
-              Users
-            </a>
+          <Link
+            href="/users"
+            className={`nav-item nav-link${
+              pathname === '/users' ? ' active' : ''
+            }`}
+          >
+            Users
           </Link>
         </div>
       </div>
@@ -50,7 +41,9 @@ export function Nav() {
           <button
             className="btn btn-outline-secondary"
             id="logout"
-            onClick={() => logout({ returnTo: window.location.origin })}
+            onClick={() =>
+              logout({ logoutParams: { returnTo: window.location.origin } })
+            }
           >
             logout
           </button>
